fine-tune previous tests, add Hero test, and a11y checks with styled-components
