+++ .eslintrc
@@ -6,8 +6,8 @@
     "cypress/globals": true
   },
   "parser": "babel-eslint",
-  "extends": ["airbnb"],
-  "plugins": ["cypress", "react-hooks"],
+  "extends": ["airbnb", "plugin:styled-components-a11y/recommended"],
+  "plugins": ["styled-components-a11y","cypress", "react-hooks"],
   "rules": {
     "cypress/no-assigning-return-values": "error",
     "cypress/no-unnecessary-waiting": "error",
