+++ src/pages/homepage.js
@@ -2,5 +2,9 @@ import React from 'react';
 import { Hero } from '../components';
 
 export default function Homepage() {
-  return <Hero />;
+  return (
+    <>
+      <Hero />
+    </>
+  );
 }
